Move `SqlTableNode` to a separate file.
