+++ metricflow/plan_conversion/to_sql_plan/dataflow_to_subquery.py
@@ -121,9 +121,9 @@ from metricflow.sql.sql_plan import (
     SqlCreateTableAsNode,
     SqlCteNode,
     SqlSelectColumn,
-    SqlTableNode,
 )
 from metricflow.sql.sql_select_node import SqlJoinDescription, SqlOrderByDescription, SqlSelectStatementNode
+from metricflow.sql.sql_table_node import SqlTableNode
 
 
 class DataflowNodeToSqlSubqueryVisitor(DataflowPlanNodeVisitor[SqlDataSet]):
