+++ metricflow/sql/sql_select_node.py
@@ -1,15 +1,16 @@
 from __future__ import annotations
 
 from dataclasses import dataclass
-from typing import Optional, Tuple, Sequence, override
+from typing import Optional, Sequence, Tuple, override
 
-from metricflow.sql.sql_plan import SqlPlanNode, SqlSelectColumn, SqlCteNode, SqlPlanNodeVisitor, SqlTableNode, \
-    SqlCteAliasMapping
 from metricflow_semantics.dag.id_prefix import IdPrefix, StaticIdPrefix
 from metricflow_semantics.dag.mf_dag import DisplayedProperty
 from metricflow_semantics.sql.sql_exprs import SqlExpressionNode
 from metricflow_semantics.sql.sql_join_type import SqlJoinType
 from metricflow_semantics.visitor import VisitorOutputT
+
+from metricflow.sql.sql_plan import SqlCteAliasMapping, SqlCteNode, SqlPlanNode, SqlPlanNodeVisitor, SqlSelectColumn
+from metricflow.sql.sql_table_node import SqlTableNode
 
 
 @dataclass(frozen=True)
