+++ tests_metricflow/sql/optimizer/test_cte_column_pruner.py
@@ -20,9 +20,9 @@ from metricflow.sql.render.sql_plan_renderer import DefaultSqlPlanRenderer, SqlP
 from metricflow.sql.sql_plan import (
     SqlCteNode,
     SqlSelectColumn,
-    SqlTableNode,
 )
 from metricflow.sql.sql_select_node import SqlJoinDescription, SqlSelectStatementNode
+from metricflow.sql.sql_table_node import SqlTableNode
 from tests_metricflow.sql.optimizer.check_optimizer import assert_optimizer_result_snapshot_equal
 
 logger = logging.getLogger(__name__)
