+++ tests_metricflow/sql/optimizer/test_table_alias_simplifier.py
@@ -16,9 +16,9 @@ from metricflow.sql.optimizer.table_alias_simplifier import SqlTableAliasSimplif
 from metricflow.sql.render.sql_plan_renderer import DefaultSqlPlanRenderer, SqlPlanRenderer
 from metricflow.sql.sql_plan import (
     SqlSelectColumn,
-    SqlTableNode,
 )
 from metricflow.sql.sql_select_node import SqlJoinDescription, SqlSelectStatementNode
+from metricflow.sql.sql_table_node import SqlTableNode
 from tests_metricflow.sql.compare_sql_plan import assert_default_rendered_sql_equal
