+++ tests_metricflow/sql/test_engine_specific_rendering.py
@@ -23,9 +23,9 @@ from metricflow_semantics.test_helpers.config_helpers import MetricFlowTestConfi
 from metricflow.protocols.sql_client import SqlClient
 from metricflow.sql.sql_plan import (
     SqlSelectColumn,
-    SqlTableNode,
 )
 from metricflow.sql.sql_select_node import SqlJoinDescription, SqlOrderByDescription, SqlSelectStatementNode
+from metricflow.sql.sql_table_node import SqlTableNode
 from tests_metricflow.sql.compare_sql_plan import assert_rendered_sql_equal
